Manage Author Version and IsCorrect Switch

+++ src/actions/editor-actions.ts
@@ -1,5 +1,6 @@
 'use server';
 
+import { auth } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 import { QuestionGroupSchema, QuestionGroupType } from "@/lib/schema";
 import { ActionResultType, ZparseOrError } from "@/lib/utils";
@@ -10,15 +11,26 @@ export const createQuestionGroup = async (data: QuestionGroupType): Promise<Acti
     const errors = ZparseOrError(QuestionGroupSchema, data)
     if (errors) return errors
 
+    const session = await auth();
+
+    if (!session) {
+        throw new Error('401 Unauthorized');
+    }
+
     try {
         const group = await prisma.questionGroup.create({
             data: {
                 name: data.name,
                 creationDate: new Date(),
+                updateDate: new Date(),
+                version: 1,
+                authorId: session.user.id!,
                 questions: {
                     create: data.questions.map(q => {
                         return {
+                            title: q.title,
                             subject: q.subject,
+                            order: q.order,
                             responses: {
                                 create: q.responses.filter(r => !!r.text).map(r => ({
                                     isCorrect: r.isCorrect,
@@ -50,6 +62,23 @@ export const updateQuestionGroup = async (data: QuestionGroupType): Promise<Acti
     const errors = ZparseOrError(QuestionGroupSchema, data)
     if (errors) return errors
 
+    const session = await auth();
+
+    if (!session) {
+        throw new Error('401 Unauthorized');
+    }
+
+    var isOwner = await prisma.questionGroup.count({
+        where: {
+            id: data.id!,
+            authorId: session.user.id!
+        }
+    })
+
+    if (isOwner == 0) {
+        throw new Error('404 Forbidden');
+    }
+
     try {
         await prisma.$transaction(async (tx) => {
 
@@ -60,6 +89,8 @@ export const updateQuestionGroup = async (data: QuestionGroupType): Promise<Acti
                 },
                 data: {
                     name: data.name,
+                    updateDate: new Date(),
+                    version: { increment: 1 }
                 },
             });
 
@@ -86,7 +117,9 @@ export const updateQuestionGroup = async (data: QuestionGroupType): Promise<Acti
                     },
                     create: {
                         groupId: data.id!,
+                        title: q.title,
                         subject: q.subject,
+                        order: q.order,
                         responses: {
                             create: q.responses.filter(r => !!r.text).map(r => ({
                                 isCorrect: r.isCorrect,
@@ -95,49 +128,53 @@ export const updateQuestionGroup = async (data: QuestionGroupType): Promise<Acti
                         }
                     },
                     update: {
+                        title: q.title,
+                        order: q.order,
                         subject: q.subject
                     }
                 });
+            }
 
-                // 4. Create / Update / Delete Responses for Updated Questions
-                const updatedQuestion = data.questions.filter(q => !!q.id)
-                for (let i = 0; i < updatedQuestion.length; i++) {
+            // 4. Create / Update / Delete Responses for Updated Questions
+            const updatedQuestion = data.questions.filter(q => !!q.id)
+            for (let i = 0; i < updatedQuestion.length; i++) {
 
-                    const q = updatedQuestion[i];
-                    // Delete
-                    await prisma.response.deleteMany({
-                        where: {
-                            id: {
-                                notIn: q.responses.filter(r => !!r.id).map(r => r.id!)
-                            },
-                            questionId: {
-                                equals: q.id!
-                            }
+                const q = updatedQuestion[i];
+                // Delete
+                await prisma.response.deleteMany({
+                    where: {
+                        id: {
+                            notIn: q.responses.filter(r => !!r.id).map(r => r.id!)
+                        },
+                        questionId: {
+                            equals: q.id!
                         }
-                    })
-
-                    for (let j = 0; j < q.responses.length; j++) {
-
-                        const r = q.responses[j];
-                        if (!r.text) continue;
-
-                        await prisma.response.upsert({
-                            where: {
-                                id: r.id || '0'
-                            },
-                            create: {
-                                questionId: q.id!,
-                                text: r.text,
-                                isCorrect: r.isCorrect
-                            },
-                            update: {
-                                text: r.text,
-                                isCorrect: r.isCorrect
-                            },
-                        });
                     }
+                })
+
+                for (let j = 0; j < q.responses.length; j++) {
+
+                    const r = q.responses[j];
+                    if (!r.text) continue;
+
+                    await prisma.response.upsert({
+                        where: {
+                            id: r.id || '0'
+                        },
+                        create: {
+                            questionId: q.id!,
+                            text: r.text,
+                            isCorrect: r.isCorrect
+                        },
+                        update: {
+                            text: r.text,
+                            isCorrect: r.isCorrect
+                        },
+                    });
                 }
             }
+        }, {
+            timeout: 5000
         })
 
         revalidatePath('/board/')
@@ -157,6 +194,23 @@ export const updateQuestionGroup = async (data: QuestionGroupType): Promise<Acti
 export const deleteQuestionGroup = async (id: string): Promise<ActionResultType<void>> => {
     const errors = ZparseOrError(z.string(), id)
     if (errors) return errors
+
+    const session = await auth();
+
+    if (!session) {
+        throw new Error('401 Unauthorized');
+    }
+
+    var isOwner = await prisma.questionGroup.count({
+        where: {
+            id: id,
+            authorId: session.user.id!
+        }
+    })
+
+    if (isOwner == 0) {
+        throw new Error('404 Forbidden');
+    }
 
     try {
         await prisma.questionGroup.delete({
